Move BackupUtil to data.backup package for better cohesion
This commit moves `BackupUtil.kt` from `eu.kanade.tachiyomi.util` to `eu.kanade.tachiyomi.data.backup`.
This utility class is tightly coupled with the backup feature and is only used by `BackupRestorer` which resides in the destination package.
Moving it improves module cohesion and cleans up the generic `util` package.

Changes:
- Moved `BackupUtil.kt` to `data/backup`
- Updated package declaration and imports in `BackupUtil.kt`
- Updated imports in `BackupRestorer.kt`
